feat: allow packages to not be signed as part of build

+++ build-package.sh
@@ -8,19 +8,24 @@ mkdir -p "${CACHE_DIR}"
 # rest of the build artifacts.
 cp "${MELANGE_CONFIG_FILE}" "${OUT_DIR}/melange.yaml"
 
-# Get a temporary file to store the private signing key
-SIGNING_KEY_FILE=$(mktemp --tmpdir="${RUNNER_TEMP}" --suffix=.rsa)
+SIGNING_ARGS=()
+if [[ "${SIGNING_KEY:-}" ]]; then
+  # Get a temporary file to store the private signing key
+  SIGNING_KEY_FILE=$(mktemp --tmpdir="${RUNNER_TEMP}" --suffix=.rsa)
 
-# Make sure we delete the private signing key file at the end of this step
-trap 'rm -f "${SIGNING_KEY_FILE}"' EXIT
+  # Make sure we delete the private signing key file at the end of this step
+  trap 'rm -f "${SIGNING_KEY_FILE}"' EXIT
 
-# Put the private signing key into a file
-chmod 600 "${SIGNING_KEY_FILE}"
-echo -n "${SIGNING_KEY}" > "${SIGNING_KEY_FILE}"
+  # Put the private signing key into a file
+  chmod 600 "${SIGNING_KEY_FILE}"
+  printf '%s' "${SIGNING_KEY}" > "${SIGNING_KEY_FILE}"
 
-# Store the public signing key in the output dir so it will be saved with
-# the rest of the build artifacts.
-openssl rsa -in "${SIGNING_KEY_FILE}" -pubout -out "${OUT_DIR}/signing.rsa.pub"
+  # Store the public signing key in the output dir so it will be saved with
+  # the rest of the build artifacts.
+  openssl rsa -in "${SIGNING_KEY_FILE}" -pubout -out "${OUT_DIR}/signing.rsa.pub"
+
+  SIGNING_ARGS=("--signing-key=${SIGNING_KEY_FILE}")
+fi
 
 # Do the actual build
 melange \
@@ -28,5 +33,5 @@ melange \
   --cache-dir="${CACHE_DIR}" \
   --out-dir="${OUT_DIR}" \
   --arch="${ARCH}" \
-  --signing-key="${SIGNING_KEY_FILE}" \
+  "${SIGNING_ARGS[@]}" \
   "${MELANGE_CONFIG_FILE}"
